generate: make record accessible from outside
New accessors for certain record fields. Enables access to those
members, planned to be used by obj_t functions.
